requirements.txt
mkl-fft==1.3.0
mkl-random==1.2.2
mkl-service==2.4.0
numpy
pandas==1.3.3
ray==1.1.0
rdkit-pypi
torch==1.7.0+cpu -f https://download.pytorch.org/whl/torch_stable.html
torch-geometric==1.6.3
torch-scatter==2.0.5 -f https://pytorch-geometric.com/whl/torch-1.7.0+cpu.html
torch-sparse==0.6.8 -f https://pytorch-geometric.com/whl/torch-1.7.0+cpu.html
tqdm==4.36.1
googledrivedownloader==0.2.0